Readme
Basic Blockchain   Prototype.

Introduction.

Blockchain is one of the   technology , which is still in its infancy  and which its potential is not fully realized yet.  Blockchain is  a distributed database of records. But what makes it unique is that it’s not a private database, but it is public and everyone who uses it has a full or partial copy of it. And a new record can be added only with a consent of the database owner.

TRANSACTION.
The data is correct, the output can be unlocked, and its value can be used to generate new outputs; if it’s not correct, the output cannot be referenced in the input. This is the mechanism that guarantees that users cannot spend coins belonging to other people.
 
BLOCK.
blockchain is just a database with certain structures: it’s an ordered, back-linked list. Which means that blocks are stored in the insertion order and that each block is linked to the previous one. This allows to quickly get the latest block in a chain and to get a block by its hash.
The structure can be implemented by using an array and a map: the array would keep ordered hashes arrays and the map would keep hash block pairs. But for our blockchain prototype we’ll just use an array, because we don’t need to get blocks by their hash for now.
create the first block of our blockchain -
 Block1 = new Block

 NODE1  NODE2 NODE3  NODE4
The blocks are store in order of insertion and latest block will be stored on the edge of blockchain.